PB620+PB636: the intrinsic argument intake — one rule, one place, three carriers
Train 44, cluster 3 (wave 43 C). Brought in from branch worktree-wf_4b186995-351-3 at
3cd83fabc. No diagnostic code used: COBOLNET2203 and COBOLNET2204 were both RETURNED.

PB620 — COMBINED-DATETIME HAS THREE CARRIERS, and the binary64 one had no argument rule at
all. `COMBINED-DATETIME(1, F)` with a COMP-2 F holding 86 400 printed 1.8639999999999 in
silence while the identical VALUE in a `PIC 9(5)V9(8)` item terminated the same run unit under
the same armed >>TURN EC-ARGUMENT-FUNCTION CHECKING ON. §15.17.3 r2 is a rule about a VALUE and
§15.5.5 bounds that form by MAGNITUDE, so the item's USAGE cannot decide whether §15.3
applies. §15.17.3's two VALUE rules now live ONLY in
CobolDate.CombinedDatetimeOutOfRange(long, CobolDec, bool); the exact body, the new
CombinedDatetimeDec (SDIDI) and CobolIntrinsics.CombinedDatetimeReal (binary64, via
CombinedDatetimeOutOfRangeReal) all ask it — one message, one raise site.
Separately the EAE: §15.17.4 r1's `a1 + (a2 / 100000)` is evaluated on the SDIDI, so under
STANDARD-DECIMAL `3661.12345678 / 100000 + 1` is 1.0366112345678 exactly, where the expression
form previously lost three digits against the literal form.
SecondsOutOfStandardForm's `(unscaled, scale)` overload now forwards to a CobolDec one: the
deleted `86 400 * 10^secScale` scaling WRAPPED past secScale 33.

PB636 — RANDOM'S SEED IS A TOTAL ARGUMENT. CobolIntrinsics.Random(long) became Random(Int128)
— the CARRIER is the totality claim — and the narrow renderer arm took IntArgWide with it, so a
19-digit seed no longer aborts the run unit with EC-ARGUMENT-FUNCTION. §15.75.3 r2's sign raise
is pinned at every width by ANegativeSeed_StillRaises_AtEveryWidth.

DETERMINATION: under NATIVE arithmetic §15.4.1 r1 does NOT bind — it is stated for the standard
modes — and COBOL.NET's approximation of COMBINED-DATETIME's EAE IS the exact/SDIDI evaluation.
Reading rejected: the note's own claim that "r1 binds under native too".

Drift tests: ExactCarrierBoundaryDriftTests.EveryArmThatCanTakeADecArgument_HasADecBody
(invariant 3, sharing ONE arm reader with invariant 2 rather than a second copy),
IntrinsicCarrierAgreementDriftTests.EveryTotalIntegerArgument_TakesTheWideIntake and
.CombinedDatetime_ScreensItsArguments_FromEveryCarrier — each with a self-test that fails on
the defect it exists for.

Goldens: 2014/pb620_combined_datetime_eae.cob (STANDARD-DECIMAL; every expected value
hand-evaluated from §15.17.4 r1), 2014/pb620_combined_datetime_carriers.cob (NATIVE — the only
mode in which the binary64 body is reachable; no binary64 VALUE is printed, because §15.4.1's
native paragraph grants an approximation and 1/100000 is not a binary fraction, so the float
legs assert only rules), 2002/pb636_random_total_seed.cob (the RULE is 85's, the WITNESS is not
writable below 2002), negative/pb620-combined-datetime-below-2014 (COBOLNET1502, verified at
85 and 2002).

FOUR DEFECTS IN THE PREDECESSOR'S UNVERIFIED CHECKPOINT DID NOT SURVIVE VERIFICATION: the new
drift test asserted 186399999990000 where §15.17.4 r1 gives the scale-10 unscaled 18 639 999
999, and compared a CobolDec by REPRESENTATION where r1 asks about VALUE (CobolDec.Compare now);
the shared screen's <summary> had been orphaned onto the new Real body; §13.18.40.3 r14 (a
31-digit data ITEM) was cited as §8.3.3.3.2 (fixed-point LITERALS) — a real clause answering a
different question, caught by audit_code_citations; and SecondsOutOfStandardForm's summary
still described the bound-scaling the fix deleted.

Rows (batches w43c-intrinsic-intake and w43c-combined-datetime-third-carrier, re-applied in
that order on the merged tree — they share AR-15.17.3-2, so they are recorded in two runs):
RV-15.4.1-1, AR-15.17.3-1, AR-15.17.3-2, RV-15.17.4-1, AR-15.75.3-2, RV-15.75.4-1,
RV-15.75.4-3 all CONFORMS. GAP 2288 -> 2288: §15 reached zero GAP before either defect was
found, so both notes carry closes_rows: [] with a reason, and the landing STRENGTHENED seven
already-OK rows whose prior test-refs each exercised only the half of the rule the defect did
not break.

+++ src/Cobol.Net.Runtime/Intrinsics/CobolIntrinsics.Float.cs
@@ -144,9 +144,27 @@ public static partial class CobolIntrinsics
     /// what NIST IF131A exercises; hazard H7). Seeds 0..32767 must yield distinct sequences (rule 3) — the .NET
     /// generator satisfies this for the whole int range. ⛔ A NEGATIVE seed violates r2 ("zero or a positive
     /// integer") and raises EC-ARGUMENT-FUNCTION (fix-queue PB65 — the old mask folded it onto a positive seed
-    /// and RANDOM(-5) silently aliased RANDOM(0x7FFFFFFB & …)); the mask remains only as the documented
+    /// and RANDOM(-5) silently aliased RANDOM(0x7FFFFFFB &amp; …)); the mask remains only as the documented
     /// wide-seed mapping for legal values beyond the generator's int range.</summary>
-    public static double Random(long seed)
+    /// <remarks>
+    /// ⛔ THE SEED IS A <b>TOTAL</b> §15 INTEGER ARGUMENT, AND THE <c>Int128</c> CARRIER IS THAT CLAIM
+    /// (kb/Work PB636 — see <c>IntrinsicRenderer.AsIntWide</c>, whose pairing guard reads this signature).
+    /// §15.75.3 r2 constrains the SIGN and nothing else — "If argument-1 is specified, it shall be zero or a
+    /// positive integer" — and §15.75.4 r3 makes the distinct-sequence subset a FLOOR, not a domain: "This
+    /// subset shall include the values from 0 through at least 32767", so a seed outside that subset is still a
+    /// legal argument whose sequence r1/r2 define. §15.3's closing paragraph raises EC-ARGUMENT-FUNCTION only on
+    /// "an incorrect value for that argument … according to the rules specified in the function definition", and
+    /// no rule here makes a large seed incorrect. The parameter was <c>long</c>, so the renderer's narrowing
+    /// intake screened the argument first: <c>FUNCTION RANDOM(S)</c> with <c>S PIC 9(19)</c> holding
+    /// 9,999,999,999,999,999,999 terminated the run unit under <c>&gt;&gt;TURN EC-ARGUMENT-FUNCTION CHECKING
+    /// ON</c> and, with checking off, substituted the ARGUMENT 0 and ran the seed-0 sequence without a word.
+    /// <para>The reduction to the generator's state is unchanged and stays the documented one
+    /// (<c>docs/CONFORMANCE.md</c> row DOC-A.1-145 under §15.75.4 r3): the low 31 bits,
+    /// <c>seed AND 0x7FFFFFFF</c>. It is the identity on 0..2³¹−1 — so the required 0..32 767 floor is injective
+    /// with a wide margin — and folds every wider seed onto that window, which r3 permits: the subset yielding
+    /// DISTINCT sequences need only include the floor, not exhaust the domain.</para>
+    /// </remarks>
+    public static double Random(Int128 seed)
     {
         if (seed < 0)
         {
